Delay pinned animation finish during app transition.
Sometimes in the case of pinned animation, we end
up with multiple animations on screen. For example when
transitioning pip->fullscreen, we are typically closing recents,
and replacing it with a task behind. We need to delay swapping ourselves
to the fullscreen stack until this app transition has finished,
otherwise we will be in the fullscreen stack with the transitioning app,
but not transitioning ourselves and thus assigned a lower Z-order.

Bug: 28333917
Bug: 27793381
Bug: 28332969
Change-Id: Ie18cc04ceeabd94093017f79676253e1189f5249