Don't dispatch Uris with appended user handle
This change fixes a bug where we started appending the user id
to the changing Uri for observers to know what user's Uri changed.
This is problematic as it changes the Uri structure expected by
observers. Also there are no public APIs to factor out the user
id from the Uri. If a system component needs to know for which
user a Uri changed it can use ContentObserver#onChange(boolean,
Uri, int) where the last argument is the user id.

bug:19312280

Change-Id: I5f0982ac3c8d6b6eab060bab65d3b53182660ac7